+++ index.js
@@ -3,8 +3,10 @@ const fs = require("fs")
 const TreeProgram = require("./src/TreeProgram.js")
 const GrammarProgram = require("./src/grammar/GrammarProgram.js")
 
-TreeProgram.executeFile = (programPath, grammarPath) =>
-  TreeProgram.makeProgram(programPath, grammarPath).execute(programPath)
+TreeProgram.executeFile = (programPath, grammarPath) => {
+  const program = TreeProgram.makeProgram(programPath, grammarPath)
+  program.execute(programPath)
+}
 
 TreeProgram.makeProgram = (programPath, grammarPath) => {
   const programClass = TreeProgram.getProgramClassFromGrammarFile(grammarPath)
@@ -17,7 +19,7 @@ TreeProgram.getProgramClassFromGrammarFile = grammarPath => {
   const grammarCode = fs.readFileSync(grammarPath, "utf8")
   const expandedGrammarCode = new TreeProgram(grammarCode).getExpanded()
   const grammarProgram = new GrammarProgram(expandedGrammarCode, grammarPath)
-  const extendedClass = grammarProgram.getParserClass() || TreeProgram
+  const extendedClass = grammarProgram.getRootParserClass() || TreeProgram
   return class extends extendedClass {
     getGrammarProgram() {
       return grammarProgram