+++ tests/jibberish/jibberishProgram.js
@@ -1,0 +1,9 @@
+const TreeProgram = require("../../index.js")
+
+class jibberishProgram extends TreeProgram {
+  executeSync() {
+    return 42
+  }
+}
+
+module.exports = jibberishProgram